Add DELETE endpoint for user deletion in UserController
This commit introduces a new HTTP DELETE endpoint in the
`UserController` class, enabling the deletion of users by
their ID. The endpoint checks for the existence of the user
and returns a "not found" response if the user does not
exist. If the user is found, it removes the user from the
database and returns a success message upon deletion.